+++ packages/repoclean/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @raulscoelho/repoclean
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @raulscoelho/script-utils@0.4.0
+
 ## 0.2.1
 
 ### Patch Changes